fixed theme routes
