+++ .gitignore
@@ -5,3 +5,4 @@ node_modules/
 dist
 tmp
 yarn-error.log
+.DS_Store
