+++ packages/client/src/api/forum/forumAPI.ts
@@ -3,11 +3,11 @@ import { TCreateTopic, TLeaveComment, TLeaveMessage } from './types';
 import { camelToSnake, snakeToCamel } from '../utils';
 const localhost = 'http://localhost:3005';
 
-const API_URL = `${localhost}/forum/`;
+const API_URL = `${localhost}/forum`;
 export const forumAPI = () => {
   const getTopics = async () => {
     try {
-      const res = await baseFetch.get(`${API_URL}topics`);
+      const res = await baseFetch.get(`${API_URL}/topics`);
       const resCamelCase = snakeToCamel(res);
       return resCamelCase;
     } catch (error: any) {
