+++ packages/client/src/api/theme/constants.ts
@@ -1,2 +1,2 @@
 const localhost = 'http://localhost:3005';
-export const API_URL = `${localhost}/profile/theme`;
+export const API_URL = `${localhost}/theme`;
