+++ packages/client/src/routes/Main/Main.tsx
@@ -9,14 +9,14 @@ import Leaderboard from '../../pages/Leaderboard/Leaderboard';
 import Forum from '../../pages/Forum/Forum';
 import Topic from '../../pages/Topic/Topic';
 import Game from '../../pages/Game/Game';
-import startSW from '../../../startServiceWorker';
+// import startSW from '../../../startServiceWorker';
 
 export const Main = () => {
   const [render, setRender] = useState(false);
 
   useEffect(() => {
     setRender(true);
-    startSW();
+    // startSW();
   }, []);
 
   if (!render) {
