+++ packages/server/middlewares/auth.ts
@@ -1,7 +1,7 @@
 import type { RequestHandler } from 'express';
 import axios from 'axios';
 
-import UserModel from '../forum/models/userModel';
+import UserModel from '../postgres/forum/models/userModel';
 
 const AUTH_ENDPOINT = 'https://ya-praktikum.tech/api/v2/auth/user';
